Rent Out Tool, Variable link on Confirmation Page
Implemented Rent Out Tool, which creates a rental with an input user id and tool id. Sets the current date as the start date and two weeks from the current date as the due date. Tools can only be rented out if they are active and not currently rented or reserved.

Made the link on the confirmation screen a ViewBag variable so that, for example, the link for the Add Tool confirmation screen can be different from the link for the Rent Out Tool confirmation screen.